Rename text input to base
